rd-29 refactor styles
