+++ packages/robindoc/src/assets/_vars.scss
@@ -1,0 +1,5 @@
+$sm: 768px;
+$md: 1024px;
+$lg: 1180px;
+$xl: 1280px;
+$xxl: 1440px;
