+++ packages/robindoc/src/assets/base.scss
@@ -1,0 +1,228 @@
+@import "./_vars";
+
+:root {
+  --header-height: 60px;
+  scroll-padding-top: calc(var(--header-height) + 20px);
+  scroll-behavior: smooth;
+  scrollbar-color: var(--neutral400) var(--neutral200);
+
+  @media screen and (width < $sm) {
+    scroll-padding-top: calc(var(--header-height) + 60px);
+  }
+}
+
+body {
+  padding: 0;
+  margin: 0;
+  accent-color: var(--primary600, #ea580c);
+  box-sizing: border-box;
+  background-color: var(--body);
+  color: var(--neutral950);
+}
+*,
+*:before,
+*:after {
+  box-sizing: inherit;
+}
+svg {
+  display: block;
+}
+
+.body-mobile-lock {
+  @media screen and (width < $sm) {
+    overflow: hidden;
+  }
+}
+
+.body-lock {
+  overflow: hidden;
+}
+
+.theme-light {
+  color-scheme: light;
+  --body: #ffffff;
+  --neutral50: #f8fafc;
+  --neutral100: #f1f5f9;
+  --neutral200: #e2e8f0;
+  --neutral300: #cbd5e1;
+  --neutral400: #94a3b8;
+  --neutral500: #64748b;
+  --neutral600: #475569;
+  --neutral700: #334155;
+  --neutral800: #1e293b;
+  --neutral900: #0f172a;
+  --neutral950: #020617;
+  --primary50: #fff7ed;
+  --primary100: #ffedd5;
+  --primary200: #fed7aa;
+  --primary300: #fdba74;
+  --primary400: #fb923c;
+  --primary500: #f97316;
+  --primary600: #ea580c;
+  --primary700: #c2410c;
+  --primary800: #9a3412;
+  --primary900: #7c2d12;
+  --primary950: #431407;
+  --link-base: #2563eb;
+  --link-base-hovered: #1e3a8a;
+  --link-visited: #4f46e5;
+  --link-visited-hovered: #3730a3;
+
+  /* SHIKI */
+  --r-cl-1: #f9826c;
+  --r-cl-2: var(--neutral100);
+  --r-cl-3: #e1e4e8;
+  --r-cl-4: #2f363d;
+  --r-cl-5: #959da5;
+  --r-cl-6: #2188ff;
+  --r-cl-7: #dbedff;
+  --r-cl-8: #005cc5;
+  --r-cl-9: #586069;
+  --r-cl-10: #6a737d;
+  --r-cl-11: #fafbfc;
+  --r-cl-12: #159739;
+  --r-cl-13: #138934;
+  --r-cl-14: #1b1f23;
+  --r-cl-15: #d1d5da;
+  --r-cl-16: #34d05822;
+  --r-cl-17: #d73a4922;
+  --r-cl-18: #ffdf5d;
+  --r-cl-19: #ffdf5d66;
+  --r-cl-20: #28a74525;
+  --r-cl-21: #d1d5da11;
+  --r-cl-22: #24292e;
+  --r-cl-23: #0366d611;
+  --r-cl-24: #f6f8fa;
+  --r-cl-25: #0366d625;
+  --r-cl-26: #34d05840;
+  --r-cl-27: #34d05800;
+  --r-cl-28: #ffd33d33;
+  --r-cl-29: #24943e99;
+  --r-cl-30: #24943e50;
+  --r-cl-31: #e36209;
+  --r-cl-32: #5a32a3;
+  --r-cl-33: #044289;
+  --r-cl-34: #cb2431;
+  --r-cl-35: #28a745;
+  --r-cl-36: #cc2536;
+  --r-cl-37: #d7dbe0;
+  --r-cl-38: #eff2f6;
+  --r-cl-39: #1b1f234d;
+  --r-cl-40: #f9c513;
+  --r-cl-41: #444d56;
+  --r-cl-42: #e2e5e9;
+  --r-cl-43: #cce5ff;
+  --r-cl-44: #ebf0f4;
+  --r-cl-45: #e8eaed;
+  --r-cl-46: #6a737d33;
+  --r-cl-47: #959da588;
+  --r-cl-48: #959da533;
+  --r-cl-49: #959da544;
+  --r-cl-50: #0366d6;
+  --r-cl-51: #3192aa;
+  --r-cl-52: #22863a;
+  --r-cl-53: #b08800;
+  --r-cl-54: #1b7c83;
+  --r-cl-55: #dbab09;
+  --r-cl-56: #6f42c1;
+  --r-cl-57: #b31d28;
+  --r-cl-58: #032f62;
+  --r-cl-59: #b31d28;
+  --r-cl-60: #ffeef0;
+  --r-cl-61: #f0fff4;
+  --r-cl-62: #ffebda;
+}
+
+.theme-dark {
+  color-scheme: dark;
+  --body: #00020a;
+  --neutral50: #020617;
+  --neutral100: #0f172a;
+  --neutral200: #1e293b;
+  --neutral300: #334155;
+  --neutral400: #475569;
+  --neutral500: #64748b;
+  --neutral600: #94a3b8;
+  --neutral700: #cbd5e1;
+  --neutral800: #e2e8f0;
+  --neutral900: #f1f5f9;
+  --neutral950: #f8fafc;
+  --primary50: #431407;
+  --primary100: #7c2d12;
+  --primary200: #9a3412;
+  --primary300: #c2410c;
+  --primary400: #ea580c;
+  --primary500: #f97316;
+  --primary600: #fb923c;
+  --primary700: #fdba74;
+  --primary800: #fed7aa;
+  --primary900: #ffedd5;
+  --primary950: #fff7ed;
+  --link-base: #3b82f6;
+  --link-base-hovered: #93c5fd;
+  --link-visited: #6366f1;
+  --link-visited-hovered: #a5b4fc;
+
+  /* SHIKI */
+  --r-cl-1: #f9826c;
+  --r-cl-2: var(--neutral100);
+  --r-cl-3: #444d56;
+  --r-cl-4: #e1e4e8;
+  --r-cl-5: #959da5;
+  --r-cl-6: #0366d6;
+  --r-cl-7: #1d2d3e;
+  --r-cl-8: #79b8ff;
+  --r-cl-9: #d1d5da;
+  --r-cl-10: #9ba3ac;
+  --r-cl-11: #24292e;
+  --r-cl-12: #176f2c;
+  --r-cl-13: #22863a;
+  --r-cl-14: #fff;
+  --r-cl-15: #586069;
+  --r-cl-16: #28a74530;
+  --r-cl-17: #d73a4930;
+  --r-cl-18: #ffd33d44;
+  --r-cl-19: #ffd33d22;
+  --r-cl-20: #2b6a3033;
+  --r-cl-21: #58606915;
+  --r-cl-22: #e1e4e8;
+  --r-cl-23: #3392ff22;
+  --r-cl-24: #2f363d;
+  --r-cl-25: #3392ff44;
+  --r-cl-26: #17e5e650;
+  --r-cl-27: #17e5e600;
+  --r-cl-28: #c6902625;
+  --r-cl-29: #17e5e699;
+  --r-cl-30: #17e5e666;
+  --r-cl-31: #ffab70;
+  --r-cl-32: #b392f0;
+  --r-cl-33: #c8e1ff;
+  --r-cl-34: #f97583;
+  --r-cl-35: #34d058;
+  --r-cl-36: #f97583;
+  --r-cl-37: #444d56;
+  --r-cl-38: #2f363d;
+  --r-cl-39: #444d56;
+  --r-cl-40: #ffea7f;
+  --r-cl-41: #d1d5da;
+  --r-cl-42: #39414a;
+  --r-cl-43: #044289;
+  --r-cl-44: #282e34;
+  --r-cl-45: #282e34;
+  --r-cl-47: #6a737d88;
+  --r-cl-48: #6a737d33;
+  --r-cl-49: #6a737d44;
+  --r-cl-50: #79b8ff;
+  --r-cl-51: #56d4dd;
+  --r-cl-52: #85e89d;
+  --r-cl-53: #ffea7f;
+  --r-cl-54: #39c5cf;
+  --r-cl-55: #ffea7f;
+  --r-cl-56: #b392f0;
+  --r-cl-57: #fdaeba;
+  --r-cl-58: #dbedff;
+  --r-cl-59: #fdaeb7;
+  --r-cl-60: #86181d;
+  --r-cl-61: #144620;
+  --r-cl-62: #c24e00;
+}
