+++ packages/robindoc/src/assets/index.ts
@@ -1,0 +1,1 @@
+import "./base.scss";
