+++ packages/robindoc/src/components/container/container.scss
@@ -1,0 +1,9 @@
+@import "../../assets/vars";
+
+.r-container {
+  max-width: $lg;
+  padding-right: 12px;
+  padding-left: 12px;
+  margin-right: auto;
+  margin-left: auto;
+}
