+++ packages/robindoc/src/components/header-social/header-social.scss
@@ -1,0 +1,20 @@
+@import "../../assets/vars";
+
+.r-header-social {
+  display: flex;
+  flex: 1;
+  justify-content: flex-end;
+
+  @media screen and (width >= $md) {
+    flex: none;
+    justify-content: unset;
+  }
+}
+
+.r-header-social-git {
+  color: var(--neutral700);
+
+  &:hover {
+    color: var(--neutral950);
+  }
+}
