+++ packages/robindoc/src/components/header/header.scss
@@ -1,0 +1,82 @@
+@import "../../assets/vars";
+
+.r-header {
+  position: sticky;
+  background-color: var(--body);
+  top: 0;
+  border-bottom: 1px solid var(--neutral300);
+  height: 60px;
+  z-index: 1000;
+}
+
+.r-header-body {
+  height: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 24px;
+}
+
+.r-header-logo {
+  display: block;
+  margin-right: 24px;
+  color: var(--neutral700);
+
+  &:hover {
+    color: var(--neutral950);
+  }
+}
+
+.r-header-nav {
+  display: flex;
+  flex-direction: column;
+  gap: 12px 16px;
+
+  @media screen and (width >= $md) {
+    flex-direction: row;
+  }
+}
+
+.r-header-link {
+  color: var(--neutral600);
+  text-decoration: none;
+  font-weight: 600;
+  padding: 4px 0;
+
+  &:hover {
+    color: var(--neutral950);
+  }
+}
+
+.r-header-actions {
+  display: flex;
+  flex-direction: column-reverse;
+  gap: 16px;
+  align-items: stretch;
+
+  @media screen and (width >= $md) {
+    flex-direction: row;
+    align-items: center;
+    gap: 28px;
+  }
+}
+
+.r-header-sections {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+
+  & .r-dropdown {
+    @media screen and (width < $sm) {
+      flex: 1;
+    }
+  }
+
+  & .r-dropdown-drop {
+    @media screen and (width < $sm) {
+      max-width: none;
+      width: 100%;
+      left: 0;
+    }
+  }
+}
