+++ packages/robindoc/src/components/heading/heading.scss
@@ -1,0 +1,31 @@
+@import "../../assets/vars";
+
+.r-h1 {
+  font-size: 40px;
+  line-height: 1.4;
+}
+
+.r-h2 {
+  font-size: 28px;
+  line-height: 1.4;
+}
+
+.r-h3 {
+  font-size: 24px;
+  line-height: 1.4;
+}
+
+.r-h4 {
+  font-size: 20px;
+  line-height: 1.4;
+}
+
+.r-h5 {
+  font-size: 18px;
+  line-height: 1.4;
+}
+
+.r-h6 {
+  font-size: 16px;
+  line-height: 1.4;
+}
