--- packages/robindoc/src/components/main.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const Main: React.FC<React.PropsWithChildren> = ({ children }) => {
-    return <main className="r-main r-container">{children}</main>;
-};
