+++ packages/robindoc/src/components/main/index.tsx
@@ -1,0 +1,7 @@
+import "./main.scss";
+import React from "react";
+import { Container } from "../container";
+
+export const Main: React.FC<React.PropsWithChildren> = ({ children }) => {
+    return <Container className="r-main">{children}</Container>;
+};
