+++ packages/robindoc/src/components/main/main.scss
@@ -1,0 +1,20 @@
+@import "../../assets/vars";
+
+.r-main {
+  min-height: calc(100vh - var(--header-height));
+  overflow-wrap: break-word;
+
+  @media screen and (width >= $sm) {
+    display: grid;
+    justify-content: center;
+    column-gap: 20px;
+    grid-template-areas: "sidebar contents" "sidebar content";
+    grid-template-columns: auto calc(100% - 232px);
+    grid-template-rows: auto 1fr;
+  }
+
+  @media screen and (width >= $lg) {
+    grid-template-areas: "sidebar content contents";
+    grid-template-columns: auto 672px auto;
+  }
+}
