+++ packages/robindoc/src/components/sidebar/sidebar.scss
@@ -1,0 +1,176 @@
+@import "../../assets/vars";
+
+.r-sidebar {
+  width: 220px;
+  grid-area: sidebar;
+
+  @media screen and (width < $sm) {
+    position: fixed;
+    width: 100%;
+    top: var(--header-height);
+    left: 0;
+    background-color: var(--body);
+    border-bottom: 1px solid var(--neutral200);
+    z-index: 10;
+    opacity: 0.98;
+    padding: 0;
+    overflow: auto;
+  }
+}
+
+.r-sidebar:not(._opened) .r-sidebar-nav {
+  @media screen and (width < $sm) {
+    display: none;
+  }
+}
+
+.r-sidebar._opened {
+  @media screen and (width < $sm) {
+    bottom: 0;
+  }
+}
+
+.r-sidebar-nav {
+  padding: 32px 12px;
+
+  @media screen and (width < $sm) {
+    padding: 12px;
+  }
+
+  @media screen and (width >= $sm) {
+    position: sticky;
+    top: var(--header-height);
+    overflow: auto;
+    scrollbar-width: thin;
+    scrollbar-color: var(--neutral200) transparent;
+    scrollbar-gutter: stable;
+    max-height: calc(100vh - var(--header-height));
+  }
+}
+
+.r-sidebar-space {
+  @media screen and (width < $sm) {
+    display: block;
+    height: 40px;
+  }
+}
+
+.r-sidebar-btn {
+  position: sticky;
+  top: 0;
+  display: flex;
+  width: 100%;
+  font-size: 14px;
+  border: 0;
+  padding: 12px;
+  background: none;
+  cursor: pointer;
+  color: var(--neutral800);
+  background-color: var(--body);
+  z-index: 1;
+
+  @media screen and (width >= $sm) {
+    display: none;
+  }
+}
+
+.r-sidebar-chevron {
+  display: inline;
+  margin-right: 6px;
+  transition: transform 0.2s;
+}
+
+.r-sidebar._opened .r-sidebar-chevron {
+  transform: rotate(90deg);
+}
+
+.r-sidebar-list {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+  font-size: 14px;
+}
+
+.r-sidebar-sublist {
+  border-left: 2px solid var(--neutral200);
+  margin-left: 12px;
+  margin-top: 2px;
+}
+
+.r-sidebar-sublist .r-sidebar-link {
+  margin-left: -2px;
+  border-left: 2px solid transparent;
+  border-top-left-radius: 0;
+  border-bottom-left-radius: 0;
+}
+
+.r-sidebar-li {
+  position: relative;
+  padding-top: 2px;
+  padding-bottom: 2px;
+  line-height: 1.125rem;
+}
+
+.r-sidebar-li-d0 {
+  margin-left: -8px;
+}
+
+.r-sidebar-p {
+  margin: 0;
+  padding: 6px 8px;
+}
+
+.r-sidebar-heading {
+  font-weight: 600;
+}
+
+.r-sidebar-link {
+  display: block;
+  padding: 6px 8px;
+  color: var(--neutral700);
+  border-radius: 4px;
+  text-decoration: none;
+}
+
+.r-sidebar-link._target {
+  color: var(--primary950);
+  background-color: var(--neutral50);
+  border-color: var(--primary800);
+}
+
+.r-sidebar-link:hover {
+  color: var(--primary800);
+  border-color: var(--primary800);
+  background-color: var(--neutral50);
+}
+
+.r-sidebar-link._active {
+  color: var(--primary700);
+  background-color: var(--neutral50);
+  border-color: var(--primary800);
+}
+
+.r-sidebar-drop-btn {
+  position: absolute;
+  right: 3px;
+  top: 5px;
+  border-radius: 4px;
+  padding: 4px;
+  background-color: var(--neutral200);
+  list-style: none;
+  cursor: pointer;
+
+  &:hover {
+    background-color: var(--neutral300);
+    color: var(--primary500);
+  }
+}
+
+.r-sidebar-drop-icon {
+  display: block;
+  transition: transform 0.3s;
+}
+
+.r-sidebar-drop[open] > .r-sidebar-drop-btn > .r-sidebar-drop-icon {
+  transform: rotate(90deg);
+}
