+++ packages/robindoc/src/components/theme/theme.scss
@@ -1,0 +1,71 @@
+@import "../../assets/vars";
+
+.r-theme {
+  position: relative;
+  display: flex;
+  gap: 4px;
+  width: 108px;
+  height: 36px;
+  padding: 2px;
+  background-color: var(--neutral950);
+  border-radius: 14px;
+
+  &::before {
+    content: "";
+    position: absolute;
+    display: block;
+    height: 32px;
+    width: 32px;
+    top: 2px;
+    left: 2px;
+    background-color: var(--body);
+    border-radius: 12px;
+    transition: left 0.2s;
+  }
+}
+
+.r-theme-btn {
+  padding: 8px;
+  background: none;
+  border: 0;
+  border-radius: 12px;
+  cursor: pointer;
+  color: var(--neutral50);
+  fill: var(--neutral950);
+  z-index: 1;
+  transition:
+    color 0.2s 0.1s,
+    fill 0.2s;
+
+  &:hover {
+    color: var(--neutral50);
+    fill: var(--neutral50);
+  }
+}
+
+.theme-dark:not(.theme-system) .r-theme::before {
+  left: 2px;
+}
+
+.theme-dark:not(.theme-system) .r-theme-btn__dark {
+  color: var(--neutral950);
+  pointer-events: none;
+}
+
+.theme-system .r-theme::before {
+  left: 38px;
+}
+
+.theme-system .r-theme-btn__system {
+  color: var(--neutral950);
+  pointer-events: none;
+}
+
+.theme-light:not(.theme-system) .r-theme::before {
+  left: 74px;
+}
+
+.theme-light:not(.theme-system) .r-theme-btn__light {
+  color: var(--neutral950);
+  pointer-events: none;
+}
